refactoring idTelegram lond date type and adding ops checker

+++ src/main/java/org/users/TelegramAdminContentUser.java
@@ -15,7 +15,7 @@ public class TelegramAdminContentUser extends TelegramOperationalUser {
     public TelegramAdminContentUser() {
     }
 
-    public TelegramAdminContentUser(String email, int idTelegramUser, String displayName) {
+    public TelegramAdminContentUser(String email, long idTelegramUser, String displayName) {
         super(email, idTelegramUser, displayName);
     }
